requirements/common.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/common.txt requirements/common.in
#
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
click==7.1.2
    # via -r requirements/common.in
idna==2.10
    # via requests
pillow==8.1.2
    # via -r requirements/common.in
requests==2.25.1
    # via -r requirements/common.in
urllib3==1.26.4
    # via requests
